pyproject.toml
[project]
name = "recorder"
version = "0.1.0"
description = "Windows screen video recorder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "imageio-ffmpeg>=0.6.0",
    "mss>=10.1.0",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
]

[project.scripts]
recorder = "recorder.main:main"
recorder-cli = "recorder.cli:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recorder"]
